Pull out user functions
